Refactoring and Sonar issues Fixes
